sprint2/step2 - dir service renamed to services
